refactor io to package; refactor docs to simplify package prefix
