+++ allel/io/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+# flake8: noqa
+from __future__ import absolute_import, print_function, division
+
+
+from .vcf_read import *
+from .vcf_write import *
+from .gff import *
+from .fasta import *
+from .util import *
