+++ docs/model/util.rst
@@ -3,7 +3,7 @@ Utility functions
 
 .. automodule:: allel.model.util
 
-.. autofunction:: create_allele_mapping
-.. autofunction:: locate_fixed_differences
-.. autofunction:: locate_private_alleles
-.. autofunction:: sample_to_haplotype_selection
+.. autofunction:: allel.create_allele_mapping
+.. autofunction:: allel.locate_fixed_differences
+.. autofunction:: allel.locate_private_alleles
+.. autofunction:: allel.sample_to_haplotype_selection
