+++ docs/stats/admixture.rst
@@ -2,10 +2,10 @@ Admixture
 =========
 
 .. automodule:: allel.stats.admixture
-.. autofunction:: patterson_f2
-.. autofunction:: patterson_f3
-.. autofunction:: patterson_d
-.. autofunction:: average_patterson_f3
-.. autofunction:: average_patterson_d
-.. autofunction:: moving_patterson_f3
-.. autofunction:: moving_patterson_d
+.. autofunction:: allel.patterson_f2
+.. autofunction:: allel.patterson_f3
+.. autofunction:: allel.patterson_d
+.. autofunction:: allel.average_patterson_f3
+.. autofunction:: allel.average_patterson_d
+.. autofunction:: allel.moving_patterson_f3
+.. autofunction:: allel.moving_patterson_d
