+++ docs/stats/distance.rst
@@ -2,10 +2,10 @@ Pairwise distance and ordination
 ================================
 
 .. automodule:: allel.stats.distance
-.. autofunction:: pairwise_distance
-.. autofunction:: plot_pairwise_distance
-.. autofunction:: pairwise_dxy
-.. autofunction:: pcoa
-.. autofunction:: condensed_coords
-.. autofunction:: condensed_coords_within
-.. autofunction:: condensed_coords_between
+.. autofunction:: allel.pairwise_distance
+.. autofunction:: allel.plot_pairwise_distance
+.. autofunction:: allel.pairwise_dxy
+.. autofunction:: allel.pcoa
+.. autofunction:: allel.condensed_coords
+.. autofunction:: allel.condensed_coords_within
+.. autofunction:: allel.condensed_coords_between
