+++ docs/stats/diversity.rst
@@ -2,15 +2,15 @@ Diversity & divergence
 ======================
 
 .. automodule:: allel.stats.diversity
-.. autofunction:: mean_pairwise_difference
-.. autofunction:: sequence_diversity
-.. autofunction:: windowed_diversity
-.. autofunction:: mean_pairwise_difference_between
-.. autofunction:: sequence_divergence
-.. autofunction:: windowed_divergence
-.. autofunction:: watterson_theta
-.. autofunction:: windowed_watterson_theta
-.. autofunction:: tajima_d
-.. autofunction:: windowed_tajima_d
-.. autofunction:: moving_tajima_d
-.. autofunction:: windowed_df
+.. autofunction:: allel.mean_pairwise_difference
+.. autofunction:: allel.sequence_diversity
+.. autofunction:: allel.windowed_diversity
+.. autofunction:: allel.mean_pairwise_difference_between
+.. autofunction:: allel.sequence_divergence
+.. autofunction:: allel.windowed_divergence
+.. autofunction:: allel.watterson_theta
+.. autofunction:: allel.windowed_watterson_theta
+.. autofunction:: allel.tajima_d
+.. autofunction:: allel.windowed_tajima_d
+.. autofunction:: allel.moving_tajima_d
+.. autofunction:: allel.windowed_df
