+++ docs/stats/fst.rst
@@ -2,15 +2,15 @@ F-statistics
 ============
 
 .. automodule:: allel.stats.fst
-.. autofunction:: weir_cockerham_fst
-.. autofunction:: hudson_fst
-.. autofunction:: patterson_fst
-.. autofunction:: average_weir_cockerham_fst
-.. autofunction:: average_hudson_fst
-.. autofunction:: average_patterson_fst
-.. autofunction:: moving_weir_cockerham_fst
-.. autofunction:: moving_hudson_fst
-.. autofunction:: moving_patterson_fst
-.. autofunction:: windowed_weir_cockerham_fst
-.. autofunction:: windowed_hudson_fst
-.. autofunction:: windowed_patterson_fst
+.. autofunction:: allel.weir_cockerham_fst
+.. autofunction:: allel.hudson_fst
+.. autofunction:: allel.patterson_fst
+.. autofunction:: allel.average_weir_cockerham_fst
+.. autofunction:: allel.average_hudson_fst
+.. autofunction:: allel.average_patterson_fst
+.. autofunction:: allel.moving_weir_cockerham_fst
+.. autofunction:: allel.moving_hudson_fst
+.. autofunction:: allel.moving_patterson_fst
+.. autofunction:: allel.windowed_weir_cockerham_fst
+.. autofunction:: allel.windowed_hudson_fst
+.. autofunction:: allel.windowed_patterson_fst
