+++ docs/stats/ld.rst
@@ -2,8 +2,8 @@ Linkage disequilibrium
 ======================
 
 .. automodule:: allel.stats.ld
-.. autofunction:: rogers_huff_r
-.. autofunction:: rogers_huff_r_between
-.. autofunction:: windowed_r_squared
-.. autofunction:: locate_unlinked
-.. autofunction:: plot_pairwise_ld
+.. autofunction:: allel.rogers_huff_r
+.. autofunction:: allel.rogers_huff_r_between
+.. autofunction:: allel.windowed_r_squared
+.. autofunction:: allel.locate_unlinked
+.. autofunction:: allel.plot_pairwise_ld
