+++ docs/stats/mendel.rst
@@ -2,8 +2,8 @@ Mendelian inheritance
 =====================
 
 .. automodule:: allel.stats.mendel
-.. autofunction:: mendel_errors
-.. autofunction:: paint_transmission
-.. autofunction:: phase_by_transmission
-.. autofunction:: phase_progeny_by_transmission
-.. autofunction:: phase_parents_by_transmission
+.. autofunction:: allel.mendel_errors
+.. autofunction:: allel.paint_transmission
+.. autofunction:: allel.phase_by_transmission
+.. autofunction:: allel.phase_progeny_by_transmission
+.. autofunction:: allel.phase_parents_by_transmission
