+++ docs/stats/sf.rst
@@ -2,25 +2,25 @@ Site frequency spectra
 ======================
 
 .. automodule:: allel.stats.sf
-.. autofunction:: sfs
-.. autofunction:: sfs_folded
-.. autofunction:: sfs_scaled
-.. autofunction:: sfs_folded_scaled
-.. autofunction:: joint_sfs
-.. autofunction:: joint_sfs_folded
-.. autofunction:: joint_sfs_scaled
-.. autofunction:: joint_sfs_folded_scaled
-.. autofunction:: fold_sfs
-.. autofunction:: fold_joint_sfs
-.. autofunction:: scale_sfs
-.. autofunction:: scale_sfs_folded
-.. autofunction:: scale_joint_sfs
-.. autofunction:: scale_joint_sfs_folded
-.. autofunction:: plot_sfs
-.. autofunction:: plot_sfs_folded
-.. autofunction:: plot_sfs_scaled
-.. autofunction:: plot_sfs_folded_scaled
-.. autofunction:: plot_joint_sfs
-.. autofunction:: plot_joint_sfs_folded
-.. autofunction:: plot_joint_sfs_scaled
-.. autofunction:: plot_joint_sfs_folded_scaled
+.. autofunction:: allel.sfs
+.. autofunction:: allel.sfs_folded
+.. autofunction:: allel.sfs_scaled
+.. autofunction:: allel.sfs_folded_scaled
+.. autofunction:: allel.joint_sfs
+.. autofunction:: allel.joint_sfs_folded
+.. autofunction:: allel.joint_sfs_scaled
+.. autofunction:: allel.joint_sfs_folded_scaled
+.. autofunction:: allel.fold_sfs
+.. autofunction:: allel.fold_joint_sfs
+.. autofunction:: allel.scale_sfs
+.. autofunction:: allel.scale_sfs_folded
+.. autofunction:: allel.scale_joint_sfs
+.. autofunction:: allel.scale_joint_sfs_folded
+.. autofunction:: allel.plot_sfs
+.. autofunction:: allel.plot_sfs_folded
+.. autofunction:: allel.plot_sfs_scaled
+.. autofunction:: allel.plot_sfs_folded_scaled
+.. autofunction:: allel.plot_joint_sfs
+.. autofunction:: allel.plot_joint_sfs_folded
+.. autofunction:: allel.plot_joint_sfs_scaled
+.. autofunction:: allel.plot_joint_sfs_folded_scaled
