+++ docs/stats/window.rst
@@ -2,8 +2,8 @@ Window utilities
 ================
 
 .. automodule:: allel.stats.window
-.. autofunction:: moving_statistic
-.. autofunction:: windowed_statistic
-.. autofunction:: windowed_count
-.. autofunction:: per_base
-.. autofunction:: equally_accessible_windows
+.. autofunction:: allel.moving_statistic
+.. autofunction:: allel.windowed_statistic
+.. autofunction:: allel.windowed_count
+.. autofunction:: allel.per_base
+.. autofunction:: allel.equally_accessible_windows
