+++ requirements_rtfd.txt
@@ -1,3 +1,4 @@
-numpy
+numpy==1.12
 numpydoc
 mock
+sphinx==1.6.2
