+++ tox.ini
@@ -29,11 +29,7 @@ deps =
 [testenv:docs]
 install_command = pip install -U -v --no-cache-dir {opts} {packages}
 basepython = python2.7
-deps =
-    sphinx==1.2
-    numpy==1.12
-    numpydoc
-    mock
+deps = -rrequirements_rtfd.txt
 commands =
     python setup.py install
     sphinx-build -W -b html -d {envtmpdir}/doctrees docs {envtmpdir}/html
